Minor improvements and bug fixes.
+ Model.getOutputData - added support for startAt parameter.  Note that
some other parameters are still not implemented (offset).

+ defined GROK.Model.STATUS enumeration.

+ Client.createProject - added support for description field.

+ added Monitor.check method to force an instantaneous invocation of the
poller, rather than needing to wait for the next interval.  Useful when
you know that the monitored entity has changed and want to update the
listeners in a uniform way.

+ improved performance of PredictionMonitor, by using startAt parameter
to start reading at the last seen row.   Possible improvement: retain
prediction contained in last row, to fill in to first row of any
subsequent data.  Without that improvement, this may be considered a
breaking change.

+fixed bug in prediction monitor that assumes that the first row is a
header row.

+ exposed GROK.Monitor type to be reusable by users of the library.